Refactor stealer into modules. Add some of the stealer API functions (as todo).
